Make Clicks to be synced in nucleus app.
It was turning out to be rather troublesome to have both app and nucleus
sync events at same time. It's certainly possible and may be easy after
this approach is done. Right now I have made it such that only one of
the apps can sync events at a time. I.e Either app or nucleus is syncing
events. I had to refactor the event sync code to be organized better
such that $window (variable used to represent on which the event
listener is to be played) has a single entry point. I should have done
it this way from the start but I instead went the easier way of simply
having objects and it gave me a hell of trouble now.
LESSON_LEARNED: Make sure your objects have single point of entry for
globals used. Always make objects constructors.