Sync scrolls across nucleus windows.
There is a small problem which essentially make this feature
useless. Browser-sync implementation of scroll sync syncs scroll on
window only, not on divs etc. So we can't sync scroll on editor window
or even sidebar. This of course need fixed.

FIX:
May be I should make a separate "NucleusSyncEvents". Or may be not. We
can use conditionals in "click" and "scroll" events as these two are the
only events that needsynced. May be form inputs as well for commit messages etc.