Refactor location event to get synced properly.
Location change of course won't sync on nucleus window as we don't have
any location there. But for the sake of consistency, I've refactored
location event as other events so it can be made accessible in nucleus
as if we need.

Once initialized, location event of course gets synced in nucleus window
as well (if we ever use iron-router in it) as it overrides .go() calls
and depend on window for back() and forward() events only.