Add code for not requiring nuclues route creation from the app.
The code is commented out though. I think we rather have user create nucleus route explicitly so she would be aware that she need to protect this route herself.

task was to have http auth system, which would take a password and would ask for that password before allowing to enter the website. I don't think it's a good way of doing stuff. Iron-router is used by most apps, I prefer hiding nucleus behind iron-router's auth which is to be set by user herself.

As per now, iron:router is a dependency, because I think it's the best
way to do it without over-committing work.