Live update Template content on changing html file
When user changes an HTML file, get the content of the changed html file
and pass it to LiveUpdate to refresh the templates; and do it in app's
window not in Nucleus' window.

TODO: Remember that we will need to execute this code in nucleus window
in future as we'll want to 100% build nucleus from within nucleus.