Done with integerating nucleuside:terminal package
Added a configuration for setting terminal's username and password.

```
Nucleus.Terminal.configure({ user: 'user', password: 'password'});
```

Not doing above on server in app will show a warning message instead of
showing a terminal. We can't allow users to access terminal with a
default user/password either, for security reasons. This terminal is too
powerful for now to risk unsafe access to it.