Show a notification on incoming video call.
Instead of just showing the video film on side, when an incoming call is
present, make the video call button glow green as a notification.

Many things about webrtc have changed in this commit. We initialize the
webrtc only when user clicks the video button. Everybody joins the same
room, so users may leave the room, stop the chat, refresh page or
whatever, the video chat is always available. All they need to do is
click the video button and they see the video film on the right.

A new system has been put in place for sending broadcast messages in
webrtc package using meteor (mongo collection). I tried a lot to get it
done with sockets of the signaling server, but it interfered with one or
another feature of the webrtc video chat. So instead I used Meteor's
collections. To make sure this won't affect performance, we publish and
subscribe to 30 messages at most and only to messages at most 5 minutes
old. This is an overkill for performance though, but this is all we
need. It adds additional benefit that we don't need to initialize webrtc
until and unless we need it. Not that initializing it had any
performance implication, but one less thing in memory is good imo.