Add normalization to added kbds, remove explicit kbd addition to Ace.
Ace was not respoinding to kbds added with keybindings plugin. So I
thought may be Ace needs explicit addition of kbds (as commands as it's
called in ace) and made many changes to add explicit kbds to Ace. But
kbds seem to working fine in Ace too after some normalization of what
user adds. I am keeping the code for adding kbds to Ace as well though,
I might be wrong and this code might be useful in future.