Add event for executing plugin code on item selection. Change the results to be given to master-prompt.
MasterPrompt requires its plugins to implement another function
`itemSelected` which is executed when user selects an item in the
prompt.

The results array used as output of the prompt is also changed. Earlier
it was an array of strings, but now it should be an array of objects
with these keys:
{
  display: "String to display in prompt",
  value: "String given to `itemSelected` as value"
}