Make LiveUpdate work for package files.
When a file belongs to a package i.e it is contained in a file in some
folder in `/packages/` directory, when it changes, new changes are
evaled in nucleus window as well as in the app window.

This brings us back to the problem of flicker in nucleus and loss of
cursor in editor which we had earlier when the files would get evaled in
nucleus.