Finish solidifying login with github.
Login with github is now solid. User can log in with github and
persistently so with cookies.

This is a breaking change though. NucleusUser was previously being used
as a system for checking who's online and who's not. Now it is a user
system. Need to change a lot of things ahead.