Check if user has access to github repo for login
Make the login more secure. We check if the `Nucleus.config.git` is a
github repo, if it is, before logging the user in, we check if the user
has access to that repo on github. Otherwise fail the login with a message.

This commit also add ability to show an Error message on failed login
instead of a black 'Unexpected Error' page